Implemented ModalCoordinator utility to indirectly communicate intent to present a modal to the modal presenter
Explored appending the modal directly to DOM but this felt wrong; a metaljs component "rendering" outside of the metaljs app is breaking scope and encapsulation. Also looked into rendering the modal directly in the component that wants to present it, but it is an unstable solution that can result in bugs caused by the app's CSS stacking context (a hidden unwanted app dependence).